requirements.txt
Pillow>=10.0.0
requests>=2.31.0
pytz>=2024.1
PyYAML>=6.0
